# KAFKA-13514: reduce consumer count and timeout to speed up the test