Reduce the timeout from 60s to 40s, and reduce the consumer count from 2000 -> 1000.
In local env, before change, the time took 44s, after change, time took 16s. 
If before change, in jenkins we took around 60s, we should now, took around 25s.

### Committer Checklist (excluded from commit message)
- [ ] Verify design and implementation 
- [ ] Verify test coverage and CI build status
- [ ] Verify documentation (including upgrade notes)