@ableegoldman 2021-12-23
@showuon  any thoughts on why this is suddenly failing? I put this test in specifically to make sure we don't introduce any nontrivial performance regressions -- my impression was that we generally had a pretty wide margin between the actual runtime and the timeout when this test was first added