@showuon 2022-01-06
@ableegoldman , yes, you're right. I didn't check the logs clearly. It looks like there's a bug in general assignor that assign the same partition to multiple consumers. I'll start to work on improving the general assignor. Thanks.